backend/requirements.txt
fastapi
uvicorn
jinja2
httpx
supabase
python-dotenv

